Feat/oleg

+++ src/main/java/ru/skypro/homework/service/impl/AdsServiceImpl.java
@@ -2,20 +2,20 @@ package ru.skypro.homework.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
-import ru.skypro.homework.dto.AdsDto;
-import ru.skypro.homework.dto.CreateAdsDto;
-import ru.skypro.homework.dto.FullAdsDto;
-import ru.skypro.homework.dto.ResponseWrapperAds;
+import ru.skypro.homework.dto.*;
 import ru.skypro.homework.entity.Ads;
 import ru.skypro.homework.entity.Image;
 import ru.skypro.homework.exceptions.AdsNotFoundException;
 import ru.skypro.homework.mapper.AdsMapper;
+import ru.skypro.homework.mapper.UserMapper;
 import ru.skypro.homework.repository.AdsRepository;
 import ru.skypro.homework.service.AdsService;
 import ru.skypro.homework.service.ImageService;
+import ru.skypro.homework.service.UserService;
 
 import java.util.List;
 
@@ -25,7 +25,9 @@ import java.util.List;
 public class AdsServiceImpl implements AdsService {
     private final AdsRepository adsRepository;
     private final AdsMapper adsMapper;
+    private final UserMapper userMapper;
     private final ImageService imageService;
+    private final UserService userService;
 
     /**
      * Receive all Ads
@@ -46,13 +48,16 @@ public class AdsServiceImpl implements AdsService {
      *
      * @param createAdsDto
      * @param image
+     * @param authentication
      * @return ad created
      */
     @Override
     @Transactional
-    public AdsDto createAds(CreateAdsDto createAdsDto, MultipartFile image) {
+    public AdsDto createAds(CreateAdsDto createAdsDto, MultipartFile image, Authentication authentication) {
         log.info("Was invoked createAds method from {}", AdsService.class.getSimpleName());
+        UserDto currentUserDto = userService.getUserByEmail(authentication.getName());
         Ads ads = adsMapper.createAdsDtoToAds(createAdsDto);
+        ads.setAuthor(userMapper.userDtoToUser(currentUserDto));
         Ads savedAds = adsRepository.save(ads);
 
         Image adsImage = imageService.createImage(image, savedAds);
@@ -68,6 +73,7 @@ public class AdsServiceImpl implements AdsService {
      */
     @Override
     public FullAdsDto getFullAdsById(long id) {
+        log.info("Was invoked getFullAdsById method from {}", AdsService.class.getSimpleName());
         Ads ads = getAdsById(id);
         return adsMapper.adsToFullAdsDto(ads);
     }
@@ -77,10 +83,13 @@ public class AdsServiceImpl implements AdsService {
      * The repository method is being used {@link AdsRepository#delete(Object)}
      *
      * @param id
+     * @param authentication
      */
     @Override
-    public void removeAds(long id) {
+    public void removeAds(long id, Authentication authentication) {
+        log.info("Was invoked removeAds method from {}", AdsService.class.getSimpleName());
         Ads ads = getAdsById(id);
+        checkIfUserHasPermission(authentication, ads);
         adsRepository.delete(ads);
     }
 
@@ -89,11 +98,14 @@ public class AdsServiceImpl implements AdsService {
      *
      * @param id
      * @param createAdsDto
+     * @param authentication
      * @return ad update
      */
     @Override
-    public AdsDto updateAdsById(long id, CreateAdsDto createAdsDto) {
+    public AdsDto updateAdsById(long id, CreateAdsDto createAdsDto, Authentication authentication) {
+        log.info("Was invoked updateAdsById method from {}", AdsService.class.getSimpleName());
         Ads oldAds = getAdsById(id);
+        checkIfUserHasPermission(authentication, oldAds);
         Ads infoToUpdate = adsMapper.createAdsDtoToAds(createAdsDto);
 
         oldAds.setPrice(infoToUpdate.getPrice());
@@ -104,6 +116,17 @@ public class AdsServiceImpl implements AdsService {
         return adsMapper.adsToAdsDto(updatedAds);
     }
 
+    private static void checkIfUserHasPermission(Authentication authentication, Ads ads) {
+        boolean matchRole = authentication.getAuthorities().stream()
+                .anyMatch(auth -> auth.getAuthority().contains(Role.ADMIN.name()));
+        boolean matchUser = authentication.getName().equals(ads.getAuthor().getEmail());
+
+        if (!matchRole || !matchUser){
+            log.warn("Current user has NO permission!");
+            throw new RuntimeException("403 Forbidden");
+        }
+    }
+
     /**
      * Receive all ads for user
      *
@@ -112,7 +135,7 @@ public class AdsServiceImpl implements AdsService {
      */
     @Override
     public ResponseWrapperAds getAllAdsForUser(String username) {
-        // TODO: 18.01.2023 Refactor with UserRepository - find user by email (thr exception), then user.getAdsList()
+        log.info("Was invoked getAllAdsForUser method from {}", AdsService.class.getSimpleName());
         List<Ads> userAdsList = adsRepository.findAdsByAuthorEmail(username);
         return adsMapper.adsListToResponseWrapperAds(userAdsList.size(), userAdsList);
     }
@@ -126,6 +149,7 @@ public class AdsServiceImpl implements AdsService {
      */
     @Override
     public Ads getAdsById(long id) {
+        log.info("Was invoked getAdsById method from {}", AdsService.class.getSimpleName());
         return adsRepository.findById(id).orElseThrow(AdsNotFoundException::new);
     }
 }
